+++ phpunit/PHPUnit/BlackBoxTest.php
@@ -5,14 +5,38 @@ namespace Tests\Innmind\BlackBox\PHPUnit;
 
 use Innmind\BlackBox\{
     PHPUnit\BlackBox,
-    PHPUnit\Framework\TestCase,
     Set,
 };
-use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\{
+    Framework\TestCase,
+    Framework\Attributes\DataProvider,
+};
 
 class BlackBoxTest extends TestCase
 {
     use BlackBox;
+
+    public function testTrait()
+    {
+        $class = new class() {
+            use BlackBox;
+
+            public function assert(): int
+            {
+                $called = 0;
+                $this
+                    ->forAll(Set\Integers::any())
+                    ->then(static function() use (&$called) {
+                        ++$called;
+                    });
+
+                return $called;
+            }
+        };
+
+        // 200 because it reads the `BLACKBOX_SET_SIZE` env var
+        $this->assertSame(200, $class->assert());
+    }
 
     public function testDoesntFailWhenTheExceptionIsExpected()
     {